requirements.txt
Flask==1.1.2
Flask-RESTful
Flask-Cors
google-cloud-storage
scikit-learn==0.22.2.post1
mysql~=0.0.2
mysql-connector-python~=8.0.21
pymysql